requirements.txt
google-adk
youtube-transcript-api
python-dotenv
fastmcp
streamlit